requirements.txt
google-cloud-storage
Flask
appengine-python-standard>=1.0.0
